+++ examples/sample_runtime/src/main.rs
@@ -26,9 +26,9 @@ use exonum::{
     node::{ApiSender, ExternalMessage, Node, NodeApiConfig, NodeChannel, NodeConfig},
     runtime::{
         rust::{BuiltinInstance, RustRuntime, ServiceFactory, Transaction},
-        AnyTx, ArtifactId, CallInfo, DeployStatus, DispatcherError, ExecutionContext,
-        ExecutionError, InstanceId, InstanceSpec, Mailbox, Runtime, SnapshotExt,
-        StateHashAggregator, WellKnownRuntime, SUPERVISOR_INSTANCE_ID,
+        AnyTx, ArtifactId, CallInfo, DispatcherError, ExecutionContext, ExecutionError, InstanceId,
+        InstanceSpec, InstanceStatus, Mailbox, Runtime, SnapshotExt, StateHashAggregator,
+        WellKnownRuntime, SUPERVISOR_INSTANCE_ID,
     },
 };
 use exonum_derive::IntoExecutionError;
@@ -334,12 +334,12 @@ fn main() {
 
         // Get an instance identifier.
         let snapshot = blockchain_ref.snapshot();
-        let (spec, status) = snapshot
+        let state = snapshot
             .for_dispatcher()
             .get_instance(instance_name.as_str())
             .unwrap();
-        assert_eq!(status, DeployStatus::Active);
-        let instance_id = spec.id;
+        assert_eq!(state.status, InstanceStatus::Active);
+        let instance_id = state.spec.id;
         // Send an update counter transaction.
         api_sender
             .broadcast_transaction(Verified::from_value(